LocalEndpointAgent/requirements.txt
grpcio>=1.66.0
grpcio-tools>=1.66.0
protobuf>=5.27.0
PyYAML>=6.0.1
psutil>=6.0.0
pydantic>=2.8.0
